workers/Worker_cross_origin_security_err.htm

<!DOCTYPE html>
<title>Worker cross-origin URL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  try {
    var w = new Worker("ftp://example.org/support/WorkerBasic.js");
    w.onerror = t.step_func_done(function(e) {
      assert_true(e instanceof Event);
    });
  } catch (e) {
    assert_unreached("Constructing a cross-origin worker should not synchronously fail. Its error event should be fired instead.");
  }
}, "Cross-origin classic workers should fail to fetch");

async_test(function(t) {
  try {
    var w = new Worker("ftp://example.org/support/WorkerBasic.js", {type: "module"});
    w.onerror = t.step_func_done(function(e) {
      assert_true(e instanceof Event);
    });
  } catch (e) {
    assert_unreached("Constructing a cross-origin module worker should not synchronously fail. Its error even should be fired instead.");
  }
}, "Cross-origin module workers should fail to fetch");
</script>
